refactor: enums 패키지 생성
