+++ src/main/java/common/enums/ContentType.java
@@ -1,12 +1,13 @@
-package common;
+package common.enums;
 
 public enum ContentType {
+    JS("text/javascript; charset=utf-8"),
+    PLAIN("text/plain; charset=utf-8"),
     HTML("text/html; charset=utf-8"),
     CSS("text/css; charset=utf-8"),
-    JS("text/javascript; charset=utf-8"),
+    ICO("image/x-icon"),
     PNG("image/png"),
     JPG("image/jpg"),
-    ICO("image/x-icon"),
     NONE(""),
     ;
 
@@ -18,5 +19,9 @@ public enum ContentType {
 
     public String getDescription() {
         return description;
+    }
+
+    public boolean isStaticContent() {
+        return this == JS || this == CSS || this == ICO || this == PNG || this == JPG;
     }
 }
