+++ src/main/java/common/enums/ResponseCode.java
@@ -1,4 +1,4 @@
-package common;
+package common.enums;
 
 public enum ResponseCode {
     OK("200 OK"),
